requirements.txt
streamlit>=1.35.0
google-generativeai>=0.7.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
python-docx>=1.1.0
pandas>=2.2.0
openpyxl>=3.1.0
numpy>=1.26.0
groq>=1.0.0
chromadb>=0.5.0
rank-bm25>=0.2.2
sentence-transformers>=3.0.0
langchain-huggingface>=0.1.0
msal>=1.28.0
requests>=2.31.0